Revert "Tricksy tests"
