--- chrome_ext/background.js
@@ -1,7 +1,0 @@
-console.log('background he he he')
-
-chrome.browserAction.onClicked.addListener(buttonClicked)
-
-function buttonClicked(tab) {
-  chrome.tabs.sendMessage(tab.id, 'hello')
-}
