--- chrome_ext/manifest.json
@@ -1,20 +1,0 @@
-{
-  "manifest_version": 2,
-  "name": "Lord of the News",
-  "version": "1.0",
-  "content_scripts": [
-    {
-      "matches": [
-        "<all_urls>"
-      ],
-      "js": [
-        "content.js"
-      ]
-    }
-  ],
-  "background": {
-    "scripts": [
-      "background.js"
-    ]
-  }
-}
