--- client/components/Animation.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-
-const Animation = () => {
-        return (
-     <div>   
-         <div class="outerCircle"></div>
-             <div class="innerCircle"></div>
-                <div class="icon"></div> 
-                <div class="center">   
-                <div class="text">News Ring</div> 
-                    <div class="ring">
-                    </div>
-                       
-                        
-                    </div>
-                </div>
-        )
-    }
-    export default Animation     
-
