+++ client/components/ArticleList.jsx
@@ -36,12 +36,15 @@ class ArticleList extends React.Component {
     );
   };
 }
+
 const mapStateToProps = state => ({
   news: state.news,
   language: state.language
 })
+
 const mapDispatchToProps = (dispatch) => {
   dispatch(fetchNews())
   return {}
 }
+
 export default connect(mapStateToProps, mapDispatchToProps)(ArticleList);
