+++ client/components/Bookmarklet.jsx
@@ -5,13 +5,7 @@ const Bookmarklet = () => {
   return (
 
     <div className="navbar navbar-expand-lg ">
-      <a href="javascript:(function () {
-  var script = document.createElement('script');
-  script.src = 'https://raw.githubusercontent.com/marie-phu-qui/Lord-of-the-News/master/server/lib/gollum.js';
-  console.log(script.src)
-  
-  document.body.appendChild(script);
-})();
+      <a href="javascript:(function () {let script = document.createElement('script'); script.src = 'noWords.js'; document.body.appendChild(script) })();
 ">Bookmarklet</a>
       {/* <p>Drag any of these links to your bookmark and gollumify or nazgulify any web page.</p> */}
     </div>
