--- tests/client/App.test.js
@@ -1,10 +1,0 @@
-// import React from 'react'
-// import { shallow } from 'enzyme'
-
-// import App from '../../client/components/App'
-
-// test('<App />', () => {
-//   const expected = 'React development has begun!'
-//   const wrapper = shallow(<App />)
-//   expect(wrapper.text()).toMatch(expected)
-// })
