+++ tests/gollum.test.js
@@ -1,13 +1,10 @@
-let { gollumify } = require('../server/lib/gollum')
-
-
-// this is the best test I'll ever write today
-test('test gollumify changing a parsed sentence to something else ?', function () {
-  let string = 'the police is on the hunt for teens kiwis'
-  let expected = 'the orc horde is on the hunt for hobbitses creatures'
+test('filterOdds should filter the odd numbers', function () {
+  // Arrange
+  let arr = [1, 2, 3, 4]
+  let expected = [1, 3]
 
   // Act
-  let actual = gollumify(string)
+  let actual = translate.filterOdds(arr)
 
   // Assert
   expect(actual).toEqual(expected)
