+++ tests/nazgul.test.js
@@ -1,15 +1,5 @@
-let { nazgulify } = require('../server/lib/nazgul')
-// import { nazgulify } from '../server/lib/nazgul'
+// let nazgulFunctions = require('../../lib/nazgul')
 
-
-
-test('test nazgulify changing a parsed sentence to something else ?', function () {
-  let string = 'Once in a blue moon'
-  let expected = 'Once in a dark moon'
-
-  // Act
-  let actual = nazgulify(string)
-
-  // Assert
-  expect(actual).toEqual(expected)
+test('is my test suite working?', function () {
+  expect(2).toEqual(2222)
 })
