--- tests/server/nazgul.test.js
@@ -1,5 +1,0 @@
-// let nazgulFunctions = require('../../../lib/nazgul')
-
-test('is my test suite working?', function () {
-  expect(2).toEqual(2222)
-})
