--- tests/setup-dom.js
@@ -1,9 +1,0 @@
-require('babel-register')
-
-const { JSDOM } = require('jsdom')
-
-const dom = new JSDOM('<body></body>')
-
-global.window = dom.window
-global.document = dom.window.document
-global.navigator = dom.window.navigator
