+++ tests/setup-env.js
@@ -1,7 +1,7 @@
-import {configure} from 'enzyme'
-import Adapter from 'enzyme-adapter-react-16'
+// import {configure} from 'enzyme'
+// import Adapter from 'enzyme-adapter-react-16'
 
 
-// set up enzyme ?
+// // set up enzyme ?
 
-configure({adapter: new Adapter()})
+// configure({adapter: new Adapter()})
